@LuciferYang 2022-07-18
- Java 8

**Before**

```
OpenJDK 64-Bit Server VM 1.8.0_322-b06 on Linux 5.13.0-1021-azure
Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz
constructor:                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
arrayOfAny                                            3              4           0       3061.3           0.3       1.0X
arrayOfAnyAsObject                                  284            290           4         35.2          28.4       0.0X
arrayOfAnyAsSeq                                      25             29           2        395.1           2.5       0.1X
arrayOfInt                                         1130           1141          16          8.8         113.0       0.0X
arrayOfIntAsObject                                 1727           1735          11          5.8         172.7       0.0X
```

**After** 

```
OpenJDK 64-Bit Server VM 1.8.0_332-b09 on Linux 5.13.0-1031-azure
Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz
constructor:                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
arrayOfAny                                            3              4           0       3049.8           0.3       1.0X
arrayOfAnyAsObject                                  302            302           0         33.2          30.2       0.0X
arrayOfAnyAsSeq                                      24             24           0        415.1           2.4       0.1X
arrayOfInt                                          728            748          21         13.7          72.8       0.0X
arrayOfIntAsObject                                 1041           1081          57          9.6         104.1       0.0X
```

- Java 11

**Before**

```
OpenJDK 64-Bit Server VM 11.0.14+9-LTS on Linux 5.13.0-1021-azure
Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
constructor:                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
arrayOfAny                                            7              7           0       1495.5           0.7       1.0X
arrayOfAnyAsObject                                    5              5           0       1845.3           0.5       1.2X
arrayOfAnyAsSeq                                       5              5           0       1845.1           0.5       1.2X
arrayOfInt                                          554            555           1         18.1          55.4       0.0X
arrayOfIntAsObject                                  766            766           0         13.1          76.6       0.0X
```

**After** 

```
OpenJDK 64-Bit Server VM 11.0.15+10-LTS on Linux 5.13.0-1031-azure
Intel(R) Xeon(R) CPU E5-2673 v3 @ 2.40GHz
constructor:                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
arrayOfAny                                            5              6           0       1880.8           0.5       1.0X
arrayOfAnyAsObject                                    5              6           1       1878.1           0.5       1.0X
arrayOfAnyAsSeq                                       5              6           0       1877.4           0.5       1.0X
arrayOfInt                                          659            673          19         15.2          65.9       0.0X
arrayOfIntAsObject                                  876            889          13         11.4          87.6       0.0X
```


- Java 17

**Before**

```
OpenJDK 64-Bit Server VM 17.0.2+8-LTS on Linux 5.13.0-1021-azure
Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz
constructor:                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
arrayOfAny                                            7              7           0       1535.9           0.7       1.0X
arrayOfAnyAsObject                                    7              7           0       1535.6           0.7       1.0X
arrayOfAnyAsSeq                                       8              8           0       1245.5           0.8       0.8X
arrayOfInt                                          581            597          12         17.2          58.1       0.0X
arrayOfIntAsObject                                  604            627          21         16.6          60.4       0.0
```

**After** 

```
OpenJDK 64-Bit Server VM 17.0.3+7-LTS on Linux 5.13.0-1031-azure
Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
constructor:                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
arrayOfAny                                            8              8           0       1245.7           0.8       1.0X
arrayOfAnyAsObject                                    8              8           0       1245.6           0.8       1.0X
arrayOfAnyAsSeq                                       8              8           0       1245.7           0.8       1.0X
arrayOfInt                                          538            539           1         18.6          53.8       0.0X
arrayOfIntAsObject                                  538            539           1         18.6          53.8       0.0X
```

I haven't update the benchmark result of scala 2.12 because there was no obvious performance change.